+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.0.5 - 2023-??-?? - ???
 
+- ZoneFileProvider, full support writing zone files out to disk
 - ZoneFileSource.list_zones added to support dynamic zone config
 
 ## v0.0.4 - 2023-05-23 - First Stop /etc/hosts